properly orient loaded xray to axes
Previously the xray was loaded as a saggital slice, but it
makes more sense to load it as a coronal slice (since it is effectively
a coronal projection). The RAS coordinates in slicer now actually
correspond to the patient coordinates in the xray, modulo any
patient rotation issues in a given xray.