start using slicer for reading file to vol node
This is a transition towards creating a volume node by using slicer's
file reading mechanism, rather than MONAI's. For the transition, Xray
objects will track a second volume node "volume_node2". In the near
future, a commit will show up that swaps out the old volume_node for
this one, and we'll go back to having a single volume node.

In this commit we start using slicer for loading volume nodes from files
and we also add a method to Xray that returns a 2D numpy array with some
nice guarantees about the image orientation.

This is work towards #7.