add dicom option for Xray loading
This is work towards #12

Currently generating segmentations is broken when loading DICOM objects.
This is because the transform that I used to overlay segmentations on
top of the original image is not taking into account the spacing aspect
of the IJK-RAS transformation.

As a start towards fixing this, this commit also removes the hard-coded
ijk-ras-directions from image_utils and instead gets the correct
directions while constructing a segmentation node by looking at the
associated volume node.