introduce eICU reader as EHR data placeholder
- introduce pandas as a python dependnency
- refactor dependency_installer into something more generic
- introduce a class Eicu for reading eICU tables and doing things with
them.

To represent loading a NICU patient, we randomly pick an ICU visit from
this dataset, and we plot and compute something for it, in this case
focusing on the FiO2 level over time from their ventilator readings.

Right now we save a plot to a file and we print out some patient
information. The goal is to next integrate these outputs into the
application instead.