improve table appearance
It seems that setting the border width of QAbstractItemView in QSS
was causing some confusing size issues with table headers.
This was difficult to debug, because the former was about setting the
style of selected items in a QComboBox popup, while the latter came up
much later when patient data tables were introduced.

close #24